feat(mission_planner): move functions to utils and add namespace
