+++ planning/autoware_mission_planner/src/lanelet2_plugins/utility_functions.hpp
@@ -29,6 +29,10 @@
 
 #include <vector>
 
+namespace autoware::mission_planner::lanelet2
+{
+using RouteSections = std::vector<autoware_planning_msgs::msg::LaneletSegment>;
+
 template <typename T>
 bool exists(const std::vector<T> & vectors, const T & item)
 {
@@ -48,4 +52,17 @@ void insert_marker_array(
 std::vector<geometry_msgs::msg::Point> convertCenterlineToPoints(const lanelet::Lanelet & lanelet);
 geometry_msgs::msg::Pose convertBasicPoint3dToPose(
   const lanelet::BasicPoint3d & point, const double lane_yaw);
+
+bool is_in_lane(const lanelet::ConstLanelet & lanelet, const lanelet::ConstPoint3d & point);
+bool is_in_parking_space(
+  const lanelet::ConstLineStrings3d & parking_spaces, const lanelet::ConstPoint3d & point);
+bool is_in_parking_lot(
+  const lanelet::ConstPolygons3d & parking_lots, const lanelet::ConstPoint3d & point);
+double project_goal_to_map(
+  const lanelet::ConstLanelet & lanelet_component, const lanelet::ConstPoint3d & goal_point);
+geometry_msgs::msg::Pose get_closest_centerline_pose(
+  const lanelet::ConstLanelets & road_lanelets, const geometry_msgs::msg::Pose & point,
+  autoware::vehicle_info_utils::VehicleInfo vehicle_info);
+
+}  // namespace autoware::mission_planner::lanelet2
 #endif  // LANELET2_PLUGINS__UTILITY_FUNCTIONS_HPP_
